Changes to GUI functions
The IGF will probably be separated in the future, but for the time being it will be built in. This is because time cannot be spent on the IGF. Also, future TODOs should add more translations. If you have something other than the current Japanese/English that can be implemented, please use the PR to suggest it.